install/deps: Factor flatNameFromTree out of install/deps
PR-URL: https://github.com/npm/npm/pull/13692/
Credit: @iarna
