+++ lib/install/deps.js
@@ -21,7 +21,7 @@ var inflateShrinkwrap = require('./inflate-shrinkwrap.js')
 var inflateBundled = require('./inflate-bundled.js')
 var andFinishTracker = require('./and-finish-tracker.js')
 var npm = require('../npm.js')
-var flatName = require('./flatten-tree.js').flatName
+var flatNameFromTree = require('./flatten-tree.js').flatNameFromTree
 var createChild = require('./node.js').create
 var resetMetadata = require('./node.js').reset
 var andIgnoreErrors = require('./and-ignore-errors.js')
@@ -466,14 +466,6 @@ var updatePhantomChildren = exports.updatePhantomChildren = function (current, c
     current.phantomChildren[moduleName(child)] = child
     current = current.parent
   }
-}
-
-function flatNameFromTree (tree) {
-  validate('O', arguments)
-  if (tree.isTop) return '/'
-  var path = flatNameFromTree(tree.parent)
-  if (path !== '/') path += '/'
-  return flatName(path, tree)
 }
 
 exports._replaceModuleByPath = replaceModuleByPath
